many changes

+++ client/src/services/post.js
@@ -1,12 +1,12 @@
 import api from './apiConfig';
 
-export const getAllPosts = async () => {
-  const resp = await api.get('/posts');
+export const getAllPosts = async (id) => {
+  const resp = await api.get(`/users/${id}/posts`);
   return resp.data;
 };
 
-export const getOnePost = async (id) => {
-  const resp = await api.get(`/posts/${id}`);
+export const getOnePost = async (user_id, id) => {
+  const resp = await api.get(`/users/${user_id}/posts/${id}`);
   return resp.data;
 };
 
@@ -15,13 +15,13 @@ export const postPost = async (postData) => {
   return resp.data;
 };
 
-export const putPost = async (id, postData) => {
-  const resp = await api.put(`/posts/${id}`, { post: postData });
+export const putPost = async (user_id, id, postData) => {
+  const resp = await api.put(`/users/${user_id}/posts/${id}`, { post: postData });
   return resp.data;
 };
 
-export const deletePost = async (id) => {
-  await api.delete(`/posts/${id}`);
+export const deletePost = async (user_id, id) => {
+  await api.delete(`users/${user_id}/posts/${id}`);
 };
 
 export const addPost = async (idData) => {
